refactor(parquet): split row filter fallback helpers
